handler for local db
