+++ splitbot.cabal
@@ -56,6 +56,7 @@ executable splitbot
                 ,  Conversation.Expense
                 ,  Core
                 ,  Currency
+                ,  LocalDB
                 ,  Queue
                 ,  Runtime
                 ,  Settings
